homework-7 v1
