+++ MortalKombat/index.html
@@ -58,6 +58,6 @@
         <div class="wall-left"></div>
         <div class="wall-right"></div>
     </div>
-    <script src="main.js"></script>
+    <script async type="module" src="main.js"></script>
 </body>
 </html>
