+++ MortalKombat/player.js
@@ -1,0 +1,29 @@
+import { changeHP, elHP, renderHP } from './utils.js';
+
+export const player1 = {
+	player: 1,
+	name: 'KITANA',
+	hp: 100,
+	img: 'http://reactmarathon-api.herokuapp.com/assets/kitana.gif',
+	weapon: ['шашка', 'нож', 'лук'],
+	attack: function(){
+		console.log(this.name + ' Fight...');
+	},
+	changeHP,
+	elHP,
+	renderHP,
+};
+
+export const player2 = {
+	player: 2,
+	name: 'SUB-ZERO',
+	hp: 100,
+	img: 'http://reactmarathon-api.herokuapp.com/assets/subzero.gif',
+	weapon: ['булава', 'меч', 'щит'],
+	attack: function(){
+		console.log(this.name + ' Fight...');
+	},
+	changeHP,
+	elHP,
+	renderHP,
+};
